fix : package rename and change component scan range
