--- domain/alarm/src/main/kotlin/com/knet/dormitory/config/JpaConfig.kt
@@ -1,8 +1,0 @@
-package com.knet.dormitory.config
-
-import org.springframework.context.annotation.Configuration
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing
-
-@Configuration
-@EnableJpaAuditing
-class JpaConfig
