--- domain/alarm/src/main/kotlin/com/knet/dormitory/config/ObjectMapperConfig.kt
@@ -1,18 +1,0 @@
-package com.knet.dormitory.config
-
-import com.fasterxml.jackson.databind.ObjectMapper
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-
-
-@Configuration
-class ObjectMapperConfig {
-
-    @Bean
-    fun objectMapper(): ObjectMapper {
-        val objectMapper = ObjectMapper()
-        objectMapper.registerModule(JavaTimeModule())
-        return objectMapper
-    }
-}
