--- domain/alarm/src/main/kotlin/com/knet/dormitory/config/WebClientConfig.kt
@@ -1,11 +1,0 @@
-package com.knet.dormitory.config
-
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import org.springframework.web.reactive.function.client.WebClient
-
-@Configuration
-class WebClientConfig {
-    @Bean
-    fun webClient():WebClient = WebClient.create()
-}
