+++ domain/notice/src/main/kotlin/com/knet/dormitory/config/JpaConfig.kt
@@ -1,8 +1,12 @@
 package com.knet.dormitory.config
 
+import org.springframework.boot.autoconfigure.domain.EntityScan
 import org.springframework.context.annotation.Configuration
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories
 
 @Configuration
 @EnableJpaAuditing
+@EnableJpaRepositories(basePackages = ["com.knet.dormitory"])
+@EntityScan(basePackages = ["com.knet.dormitory"])
 class JpaConfig
