--- web/api-server-mvc/src/main/kotlin/com/knet/dormitory/web/shared/CommonResponse.kt
@@ -1,10 +1,0 @@
-package com.knet.dormitory.web.shared
-
-import org.springframework.http.HttpStatus
-
-class CommonResponse(
-    val status: HttpStatus,
-    val message:String
-) {
-    override fun toString(): String = "CommonResponse(status=$status, message='$message')"
-}
