--- web/api-server-mvc/src/main/kotlin/com/knet/dormitory/web/shared/dto/ErrorResponse.kt
@@ -1,8 +1,0 @@
-package com.knet.dormitory.web.shared.dto
-
-import org.springframework.http.HttpStatus
-
-class ErrorResponse(
-    val status:HttpStatus,
-    val message:String
-)
