+++ web/api-server-mvc/src/main/kotlin/com/knet/mvc/web/shared/CommonResponse.kt
@@ -1,0 +1,10 @@
+package com.knet.mvc.web.shared
+
+import org.springframework.http.HttpStatus
+
+class CommonResponse(
+    val status: HttpStatus,
+    val message:String
+) {
+    override fun toString(): String = "CommonResponse(status=$status, message='$message')"
+}
