+++ web/api-server-mvc/src/main/kotlin/com/knet/mvc/web/shared/dto/ErrorResponse.kt
@@ -1,0 +1,8 @@
+package com.knet.mvc.web.shared.dto
+
+import org.springframework.http.HttpStatus
+
+class ErrorResponse(
+    val status:HttpStatus,
+    val message:String
+)
