+++ web/batch-server/src/main/kotlin/com/knet/batch/DormitoryBatchApplication.kt
@@ -1,0 +1,11 @@
+package com.knet.batch
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication(scanBasePackages = ["com.knet.dormitory"])
+class DormitoryBatchApplication
+
+fun main(args: Array<String>) {
+    runApplication<DormitoryBatchApplication>(*args)
+}
