--- web/batch-server/src/main/kotlin/com/knet/dormitory/DormitoryBatchApplication.kt
@@ -1,11 +1,0 @@
-package com.knet.dormitory
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-
-@SpringBootApplication(scanBasePackages = ["com.knet.dormitory"])
-class DormitoryBatchApplication
-
-fun main(args: Array<String>) {
-    runApplication<DormitoryBatchApplication>(*args)
-}
