--- web/batch-server/src/main/kotlin/com/knet/dormitory/config/WebClientConfig.kt
@@ -1,19 +1,0 @@
-package com.knet.dormitory.config
-
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import org.springframework.web.reactive.function.client.ExchangeStrategies
-import org.springframework.web.reactive.function.client.WebClient
-
-@Configuration
-class WebClientConfig {
-    @Bean
-    fun webClient(): WebClient {
-        val exchangeStrategies = ExchangeStrategies.builder()
-            .codecs { configurer -> configurer.defaultCodecs().maxInMemorySize(-1) }
-            .build()
-        return WebClient.builder()
-            .exchangeStrategies(exchangeStrategies)
-            .build()
-    }
-}
